moved allocator to data package
